+++ examples/storybook/stories/Components/EditorWrapper.js
@@ -147,7 +147,7 @@ const EditorWrapper = ({
   onChange,
   isMobile = false,
   pluginsToDisplay,
-  getToolbarSettings,
+  toolbarSettings,
 }) => {
   const editorPlugins = pluginsToDisplay
     ? pluginsToDisplay.map(plugin => pluginsMap[plugin])
@@ -160,7 +160,7 @@ const EditorWrapper = ({
       content={content}
       isMobile={isMobile}
       placeholder={'Share something...'}
-      getToolbarSettings={getToolbarSettings}
+      toolbarSettings={toolbarSettings}
     >
       <RichContentEditor helpers={{ onFilesChange }} onChange={onChange} />
     </RicosEditor>
@@ -173,7 +173,7 @@ EditorWrapper.propTypes = {
   onChange: PropTypes.func,
   isMobile: PropTypes.bool,
   pluginsToDisplay: PropTypes.arrayOf(PropTypes.string),
-  getToolbarSettings: PropTypes.object,
+  toolbarSettings: PropTypes.object,
 };
 
 export default EditorWrapper;